fixed bug for Command :: get
Fixed the problem that overload resolution could not be done when using floating point type

+++ example.cpp
@@ -28,7 +28,7 @@ int main(int argc, char const *argv[])
         cout << cmd.get<string>("output") << endl;
     }
     if(cmd.has("max")){
-        cout << "max is " << cmd.get<int>("max") << endl;
+        cout << "max is " << cmd.get<double>("max") << endl;
     }
 
     return 0;
